+++ src/game/client/components/blockscore.cpp
@@ -166,8 +166,8 @@ void CBlockScore::WriteTop3(int index, char *pResult, int pResultSize, void *pDa
 
 int CBlockScore::GetIndex(vec2 Pos)
 {
-	int Nx = clamp(round(Pos.x)/32, 0, m_MapWidth-1);
-	int Ny = clamp(round(Pos.y)/32, 0, m_MapHeight-1);
+	int Nx = clamp(int(round(Pos.x)/32), 0, m_MapWidth-1);
+	int Ny = clamp(int(round(Pos.y)/32), 0, m_MapHeight-1);
 	int Tile = Ny*m_MapWidth+Nx;
 	return m_pTiles[Tile].m_Index;
 }
@@ -472,8 +472,8 @@ bool CBlockScore::OnInput(IInput::CEvent e)
 		{
 			vec2 MousePosWorld = m_pClient->m_pControls->m_TargetPos;
 
-			int Nx = clamp(round(MousePosWorld.x)/32, 0, m_MapWidth-1);
-			int Ny = clamp(round(MousePosWorld.y)/32, 0, m_MapHeight-1);
+			int Nx = clamp(int(round(MousePosWorld.x)/32), 0, m_MapWidth-1);
+			int Ny = clamp(int(round(MousePosWorld.y)/32), 0, m_MapHeight-1);
 			int Tile = Ny*m_MapWidth+Nx;
 
 			m_pTiles[Tile].m_Index = m_DrawMode;
